pyproject.toml
[project]
name = "fuck-rainclassroom"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai",
    "opencv-python",
    "numpy",
    "requests",
    "paddleocr",
    "paddlepaddle",
    "websocket-client",
    "DrissionPage",
]
